お気に入り処理

+++ app/controllers/favorites_controller.rb
@@ -7,14 +7,18 @@ class FavoritesController < ApplicationController
   end
 
   def destroy
-    @clothing = Clothing.find(params[:clothing_id])
-    favorite = Favorite.find_by(clothing_id: params[:clothing_id], user_id: current_user.id)
-    favorite.destroy
-    redirect_to clothing_path(@clothing.id)
+    clothing = Clothing.find(params[:clothing_id])
+    favorite = current_user.favorites.find_by(clothing_id: clothing.id)
+
+    if favorite
+      favorite.destroy
+    end
+
+    redirect_to clothing_path(clothing.id)
   end
 
   def index
-    @clothing = Clothing.find(params[:clothing_id])
     @favorite_clothings = current_user.favorite_clothings
+
   end
 end
